Partially revert [f4f51eb](https://github.com/google/tsunami-security-scanner/commit/f4f51eb385b6242972556eaa5392617da1d4bacc).
That change accidentally removed shadowJar creation, which is heavily relied on by plugins for dependencies. In other words, the core engine propagates some dependencies for the plugins.

In the future, we want to get rid of that behavior and explicitly define dependencies for every plugin, but this will require some rework.

PiperOrigin-RevId: 802576949
Change-Id: Iafb9f4bda69c6a47066b5ed708706efe687e92e1